requirements.txt
# Core dependencies for COGS
elasticsearch==8.19.0
mistralai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Document processing
PyPDF2>=3.0.0
mammoth>=1.5.0

# Logging and monitoring
requests>=2.28.0
tqdm>=4.64.0
colorama>=0.4.4

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Web framework for future API development
flask>=2.2.0
flask-cors>=4.0.0

# Optional: Jupyter for data exploration
jupyter>=1.0.0
